Move files to lib in blocks
